+++ examples/private_cluster/modules.tf
@@ -1,7 +1,5 @@
 locals {
-
   allowed_cidrs = ["x.x.x.x", "y.y.y.y"]
-
 }
 
 module "azure_region" {
@@ -34,7 +32,6 @@ module "azure_virtual_network" {
   resource_group_name = module.rg.resource_group_name
 
   vnet_cidr = ["10.0.0.0/19"]
-
 }
 
 resource "azurerm_private_dns_zone" "private_dns_zone" {
@@ -58,7 +55,6 @@ module "node_network_subnet" {
   subnet_cidr_list = ["10.0.0.0/20"]
 
   service_endpoints = ["Microsoft.Storage"]
-
 }
 
 module "appgtw_network_subnet" {
@@ -74,7 +70,6 @@ module "appgtw_network_subnet" {
   virtual_network_name = module.azure_virtual_network.virtual_network_name
 
   subnet_cidr_list = ["10.0.20.0/24"]
-
 }
 
 module "global_run" {
@@ -93,7 +88,6 @@ module "global_run" {
   resource_group_name = module.rg.resource_group_name
 
   tenant_id = var.azure_tenant_id
-
 }
 
 module "aks" {
@@ -152,10 +146,9 @@ module "aks" {
   oms_log_analytics_workspace_id = module.global_run.log_analytics_workspace_id
   azure_policy_enabled           = false
 
+  logs_destinations_ids = [module.global_run.log_analytics_workspace_id]
 
-  diagnostic_settings_logs_destination_ids = [module.global_run.log_analytics_workspace_id]
-
-  appgw_ingress_controller_values = { "verbosityLevel" = "5", "appgw.shared" = "true" }
+  appgw_ingress_controller_values = { "verbosityLevel" = 5, "appgw.shared" = true }
   cert_manager_settings           = { "cainjector.nodeSelector.agentpool" = "default", "nodeSelector.agentpool" = "default", "webhook.nodeSelector.agentpool" = "default" }
   velero_storage_settings         = { allowed_cidrs = local.allowed_cidrs }