+++ r-infra.tf
@@ -13,5 +13,4 @@ module "infra" {
   aadpodidentity_custom_name      = var.aadpodidentity_custom_name
 
   aadpodidentity_extra_tags = var.aadpodidentity_extra_tags
-
 }