+++ r-aks.tf
@@ -1,5 +1,5 @@
+#tfsec:ignore:azure-container-use-rbac-permissions
 resource "azurerm_kubernetes_cluster" "aks" {
-
   name                            = local.aks_name
   location                        = var.location
   resource_group_name             = var.resource_group_name